+++ src/test/java/com/oheat/food/jpaTest/ShopRepositoryTest.java
@@ -9,6 +9,7 @@ import com.oheat.food.entity.ShopJpaEntity;
 import com.oheat.food.repository.CategoryJpaRepository;
 import com.oheat.food.repository.ShopJpaRepository;
 import java.util.List;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -97,5 +98,53 @@ public class ShopRepositoryTest {
         List<ShopJpaEntity> result = shopJpaRepository.findByCategory(category);
 
         assertThat(result.size()).isEqualTo(3);
+    }
+
+    @Test
+    @DisplayName("치킨 매장이 7개이고 5개 단위로 페이징할 때, 조회 결과는 0번 페이지는 5개, 1번 페이지는 2개이다")
+    void givenSevenShops_whenFindShopByCategory_thenPage0Return5AndPage1Return2() {
+
+    }
+
+    @Disabled
+    @Test
+    @DisplayName("카테고리로 매장 목록 조회 시 정렬 기준이 기본순이면, 최근 등록 순으로 조회")
+    void whenSortByDefault_thenRecentRegistrationOrder() {
+
+    }
+
+    @Disabled
+    @Test
+    @DisplayName("카테고리로 매장 목록 조회 시 정렬 기준이 배달팁 낮은 순이면, 배달팁 오름차순으로 조회")
+    void whenSortByDeliveryTip_thenDeliveryTipAscendingOrder() {
+
+    }
+
+    @Disabled
+    @Test
+    @DisplayName("카테고리로 매장 목록 조회 시 정렬 기준이 최소주문 금액 낮은 순이면, 최소주문 금액 오름차순으로 조회")
+    void whenSortByMinimumOrderAmount_thenMinimumOrderAmountAscendingOrder() {
+
+    }
+
+    @Disabled
+    @Test
+    @DisplayName("카테고리로 매장 목록 조회 시 정렬 기준이 가까운 순이면, 내 위치에서의 거리 오름차순으로 조회")
+    void whenSortByNearestOrder_thenDistanceInMyLocationAscendingOrder() {
+
+    }
+
+    @Disabled
+    @Test
+    @DisplayName("카테고리로 매장 목록 조회 시 정렬 기준이 주문 많은 순이면, 누적 주문 수 내림차순으로 조회")
+    void whenSortByOrderAmount_thenOrderAmountDescendingOrder() {
+
+    }
+
+    @Disabled
+    @Test
+    @DisplayName("카테고리로 매장 목록 조회 시 정렬 기준이 별점 높은 순이면, 별점 내림차순으로 조회")
+    void whenSortByReviewScore_thenReviewScoreDescendingOrder() {
+
     }
 }
